Revert "test: Add utility functions for e2e tests (#393)"
This reverts commit be5637e0af1f51bf3b533ec25c5603da50b9e44a.
