+++ test/e2e/editor/shared/assets.js
@@ -1,7 +1,7 @@
 import {click} from "../../shared/util.js";
 import {getContentWindowElement} from "./contentWindows.js";
 import {clickContextMenuItem} from "./contextMenu.js";
-import {getTreeViewItemElement, waitForTreeViewDisappear} from "./treeView.js";
+import {getNotTreeViewItemElement, getTreeViewItemElement} from "./treeView.js";
 
 /**
  * Clicks the create asset button in the project window and clicks the specified
@@ -34,14 +34,22 @@ export async function createAsset(page, testContext, createMenuPath) {
 /**
  * @template {boolean} TIsNotFunction
  * @param {import("puppeteer").Page} page
+ * @param {string[]} assetPath
+ * @param {TIsNotFunction} isNotFunction
  */
-async function getProjectRootTreeViewEl(page) {
+async function getAssetTreeViewHelper(page, assetPath, isNotFunction) {
 	const projectEl = await getContentWindowElement(page, "project");
 	const projectRootTreeViewEl = await projectEl.$(":scope > .editorContentWindowContent > .treeViewItem");
 	if (!projectRootTreeViewEl) {
 		throw new Error("Project root treeview element not found.");
 	}
-	return projectRootTreeViewEl;
+	let result;
+	if (isNotFunction) {
+		result = await getNotTreeViewItemElement(page, projectRootTreeViewEl, assetPath);
+	} else {
+		result = await getTreeViewItemElement(page, projectRootTreeViewEl, assetPath);
+	}
+	return /** @type {TIsNotFunction extends true ? import("puppeteer").JSHandle<boolean> : import("puppeteer").ElementHandle} */ (result);
 }
 
 /**
@@ -50,19 +58,16 @@ async function getProjectRootTreeViewEl(page) {
  * @param {string[]} assetPath
  */
 export async function getAssetTreeView(page, assetPath) {
-	const projectRootTreeViewEl = await getProjectRootTreeViewEl(page);
-	const result = await getTreeViewItemElement(page, projectRootTreeViewEl, assetPath);
-	return /** @type {import("puppeteer").ElementHandle} */ (result);
+	return await getAssetTreeViewHelper(page, assetPath, false);
 }
 
 /**
- * Finds the project window and waits for an asset treeview to disappear.
+ * Finds the project window and searches for the specified asset treeview element.
  * @param {import("puppeteer").Page} page
  * @param {string[]} assetPath
  */
-export async function waitForAssetDissappear(page, assetPath) {
-	const projectRootTreeViewEl = await getProjectRootTreeViewEl(page);
-	await waitForTreeViewDisappear(page, projectRootTreeViewEl, assetPath);
+export async function getNotAssetTreeView(page, assetPath) {
+	return await getAssetTreeViewHelper(page, assetPath, true);
 }
 
 /**
